Part3/requirements/testing.txt
# 测试环境依赖项，在生产环境基础上增加用于测试代码的包
-r production.txt

pytest
coverage
pytest-cov
pytest-flask
